feat(ios): close chat → SwiftData expense-logging loop
Tapping "Confirmar" on a chat expense card now writes a real
Transaction row to the SwiftData store instead of just flipping a
UI flag. Home's @Query picks it up automatically and the available
balance updates without a refresh.

Three changes ride together because they touch the same surface
and shipping them separately would mean half-finished state:

1. ExpenseProposal.amount: Double → Decimal (Constitution P3).
   Updated emit sites in ChatPromptKit, RuleBasedBackend, and the
   display in ChatBubbleRow (now uses CurrencyFormatter.format).
   AppleFoundationModelsBackend consumes ChatPromptKit's parser so
   it inherits the precision fix transparently.

2. ExpenseQuickParser (new, in Helpers/) lifts the merchant+amount
   regex out of RuleBasedBackend. ChatViewModel.send() runs it
   before any backend so "iFood 47" / "Uber 23,50" produces an
   expense card with no model warmup or typing dot. RuleBasedBackend
   now delegates to the same parser; the LLM tiers fall through
   only for free-form text they can actually help with.

3. ChatViewModel.confirmExpense extracts the proposal from the
   matching thread bubble, coerces the amount to a negative Decimal
   (Transaction.swift sign convention: positive = income), and
   inserts via the model context that ChatView pushes in from
   @Environment(\.modelContext) on appear.

Tests (ChatExpenseLoopTests, four cases):
- round-trip from send("iFood 47") through confirmExpense to a
  stored Transaction with the right amount, sign, and category
- quick-parse short-circuits the rules backend (no typing dot,
  no chip row appended after the user message)
- confirmExpense on a non-expense bubble is a safe no-op
- non-expense inputs ("como tô esse mês?", "oi") don't match the
  quick parser and pass through to the backend

Also adds ExpenseProposal.proposedDate() — resolves "Hoje" /
"Ontem" / "dd/MM" plus "HH:mm" into a concrete Date for the
Transaction timestamp.

+++ ios/Due/Due/Views/Chat/ChatModels.swift
@@ -17,12 +17,16 @@ struct ChatMessage: Identifiable {
     }
 }
 
+// Constitution Principle 3: currency is Decimal end-to-end. The chat
+// surface (rules / Llama JSON / FoundationModels) was using Double; this
+// is now the source of truth and every emit/consume site is responsible
+// for staying in Decimal.
 struct ExpenseProposal {
     let merchant: String
-    let amount: Double
+    let amount: Decimal
     let category: String
-    let date: String
-    let time: String
+    let date: String      // "Hoje" / "Ontem" / "dd/MM"  — see proposedDate()
+    let time: String      // "HH:mm"
 }
 
 struct ChatInsight {
